requirements.txt
streamlit
pandas
python-dotenv
sentence-transformers
langchain-google-genai
langchain-core
langchain
chromadb
pydantic